spanwCount per frame

+++ DXEngine/Particle.h
@@ -13,7 +13,7 @@ namespace DE {
 	struct ParticleConsts {
 		float dt;              // 델타 타임
 		float time;            // 경과 시간 (랜덤 시드용)
-		float spawnCount;       // 이번 프레임에 생성할 개수
+		UINT spawnCount;       // 이번 프레임에 생성할 개수
 		UINT maxParticles;     // 최대 파티클 수
 	};
 }
